refactor config utils to prevent circular import
